+++ CHANGELOG.md
@@ -15,6 +15,16 @@ You can also refer to commit logs to get details on what was implemented, fixed 
 - Add `UTF8DecodeError`
 
 
+### 2.1.2
+
+- Fix issue #75 (refresh tokens not saved to keychain)
+
+
+### 2.1.1
+
+- Fix issue #72 (refresh token SNAFU from 2.1)
+
+
 ### 2.1
 
 - Refresh tokens now work for all grants
@@ -22,6 +32,7 @@ You can also refer to commit logs to get details on what was implemented, fixed 
 - Improvements to password grant
 - Properly implement dynamic client registration
 - Fix issues #47, #59, #61, #66 and improve behavior in several scenarios
+
 
 ### 2.0.2